refactor(frontend): extract axios, fix type imports
