+++ frontend/src/components/FileList.tsx
@@ -14,8 +14,8 @@ import MuiTypography from "@mui/material/Typography"
 import { byteSizeToUnits } from "../utils"
 import { useLocationContext } from "../contexts/LocationContext"
 import { useAsyncTaskContext } from "../contexts/AsyncTaskContext"
-import { type FileData, useFileMutations, useFileFetches } from "../hooks/files"
-
+import { useFileMutations, useFileFetches } from "../hooks/files"
+import { type FileData } from "../types/files"
 interface FileListProps {
 	data: Array<FileData>
 }
