+++ frontend/tests/FileDetails.test.tsx
@@ -5,7 +5,7 @@ import { type UseQueryResult } from "@tanstack/react-query"
 
 import { renderWithContexts as render, getAxiosMockAdapter } from "./helpers"
 import FileDetails from "../src/components/FileDetails"
-import { type FileData } from "../src/hooks/files"
+import { type FileData } from "../src/types/files"
 import * as fileQueries from "../src/hooks/files"
 import * as locationContextUtils from "../src/contexts/LocationContext"
